backend/requirements.txt
# Python Backend Requirements
flask>=3.0.0
flask-cors>=4.0.0
selenium>=4.15.0
pillow>=10.0.0
requests>=2.28.0
